requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.11
requests==2.31.0 